move protobuf files; adjust imports everywhere.
