+++ crypto/pb/crypto.proto
@@ -2,6 +2,8 @@ syntax = "proto2";
 
 package crypto.pb;
 
+option go_package = "github.com/libp2p/go-libp2p-core/crypto/pb";
+
 enum KeyType {
 	RSA = 0;
 	Ed25519 = 1;
