+++ record/envelope.go
@@ -7,7 +7,7 @@ import (
 
 	pool "github.com/libp2p/go-buffer-pool"
 	"github.com/libp2p/go-libp2p-core/crypto"
-	pb "github.com/libp2p/go-libp2p-core/crypto/pb"
+	pb "github.com/libp2p/go-libp2p-core/record/pb"
 
 	"github.com/gogo/protobuf/proto"
 	"github.com/multiformats/go-varint"
