+++ record/pb/Makefile
@@ -4,7 +4,7 @@ GO = $(PB:.proto=.pb.go)
 all: $(GO)
 
 %.pb.go: %.proto
-		protoc --proto_path=$(PWD)/../..:. --gogofaster_out=. $<
+		protoc --proto_path=$(PWD):$(PWD)/../.. --gogofaster_out=. $<
 
 clean:
 		rm -f *.pb.go
